Address remaining PR review threads with hardening and coverage
