+++ crates/uc-cli/src/commands/migrate.rs
@@ -1,0 +1,116 @@
+use super::*;
+
+pub(crate) fn run_migrate(args: MigrateArgs) -> Result<()> {
+    let manifest_path = resolve_manifest_path(&args.manifest_path)?;
+    let workspace_root = manifest_path
+        .parent()
+        .context("manifest path has no parent")?
+        .to_path_buf();
+    let raw = read_text_file_with_limit(&manifest_path, MAX_MANIFEST_BYTES, "manifest")?;
+    let parsed: TomlValue = raw
+        .parse()
+        .with_context(|| format!("failed to parse TOML in {}", manifest_path.display()))?;
+
+    let package = parsed.get("package").and_then(TomlValue::as_table);
+    let package_name = package
+        .and_then(|tbl| tbl.get("name"))
+        .and_then(TomlValue::as_str)
+        .map(str::to_string);
+    let package_version = package
+        .and_then(|tbl| tbl.get("version"))
+        .and_then(TomlValue::as_str)
+        .map(str::to_string);
+    let edition = package
+        .and_then(|tbl| tbl.get("edition"))
+        .and_then(TomlValue::as_str)
+        .map(str::to_string);
+
+    let dependency_count = parsed
+        .get("dependencies")
+        .and_then(TomlValue::as_table)
+        .map_or(0, |tbl| tbl.len());
+    let dev_dependency_count = parsed
+        .get("dev-dependencies")
+        .and_then(TomlValue::as_table)
+        .map_or(0, |tbl| tbl.len());
+
+    let known_sections = [
+        "package",
+        "dependencies",
+        "dev-dependencies",
+        "workspace",
+        "target",
+        "scripts",
+        "tool",
+        "features",
+        "patch",
+        "cairo",
+        "lib",
+        "executable",
+        "test",
+    ];
+
+    let unknown_sections = parsed
+        .as_table()
+        .map(|tbl| {
+            let mut keys: Vec<String> = tbl
+                .keys()
+                .filter(|k| !known_sections.contains(&k.as_str()))
+                .cloned()
+                .collect();
+            keys.sort();
+            keys
+        })
+        .unwrap_or_default();
+
+    let mut warnings = Vec::new();
+    if package_name.is_none() {
+        warnings.push("missing [package].name".to_string());
+    }
+    if edition.is_none() {
+        warnings.push("missing [package].edition".to_string());
+    }
+    if !unknown_sections.is_empty() {
+        warnings.push(format!(
+            "unknown top-level sections detected: {}",
+            unknown_sections.join(", ")
+        ));
+    }
+
+    let report = MigrationReport {
+        generated_at_epoch_ms: epoch_ms()?,
+        manifest_path: manifest_path.display().to_string(),
+        workspace_root: workspace_root.display().to_string(),
+        package_name: package_name.clone(),
+        package_version: package_version.clone(),
+        edition: edition.clone(),
+        dependency_count,
+        dev_dependency_count,
+        unknown_sections,
+        warnings,
+        suggested_next_steps: vec![
+            "Run `uc compare-build` to establish artifact parity before migration.".to_string(),
+            "Define migration owner and target milestone for this workspace.".to_string(),
+            "Prepare `Uc.toml` and validate in CI shadow lane.".to_string(),
+        ],
+    };
+
+    let report_path = args
+        .report_path
+        .unwrap_or_else(|| workspace_root.join("uc-migration-report.json"));
+    write_json_report(&report_path, &report)?;
+    println!("Migration report: {}", report_path.display());
+
+    if let Some(uc_toml_path) = args.emit_uc_toml {
+        write_uc_toml(
+            &uc_toml_path,
+            &manifest_path,
+            package_name.as_deref(),
+            package_version.as_deref(),
+            edition.as_deref(),
+        )?;
+        println!("Generated Uc.toml scaffold: {}", uc_toml_path.display());
+    }
+
+    Ok(())
+}
